+++ Offroad-iOS/Offroad-iOS/Presentation/Diary/MyDiaryManager/MyDiaryManager.swift
@@ -5,7 +5,7 @@
 //  Created by 조혜린 on 3/11/25.
 //
 
-import Foundation
+import UIKit
 
 import RxSwift
 import RxRelay
@@ -52,5 +52,23 @@ extension MyDiaryManager {
         case .currentPage:
             return (calendar.component(.year, from: currentPageDate), calendar.component(.month, from: currentPageDate))
         }
+    }
+    
+    func showCompleteCreateDiaryAlert(viewController: UIViewController) {
+        let alertController = ORBAlertController(title: DiaryMessage.completeCreateDiaryTitle, message: DiaryMessage.completeCreateDiaryMessage, type: .normal)
+        alertController.configureMessageLabel{ label in
+            label.setLineHeight(percentage: 150)
+        }
+        alertController.xButton.isHidden = true
+        let cancelAction = ORBAlertAction(title: "나중에", style: .cancel) { _ in return }
+        let okAction = ORBAlertAction(title: "확인", style: .default) { _ in
+            let diaryViewController = DiaryViewController(shouldShowLatestDiary: true)
+            diaryViewController.setupCustomBackButton(buttonTitle: "")
+            viewController.navigationController?.pushViewController(diaryViewController, animated: true)
+        }
+        alertController.addAction(cancelAction)
+        alertController.addAction(okAction)
+        
+        viewController.present(alertController, animated: true)
     }
 }
